instagram/requirements.txt
asgiref==3.5.0
Django==4.0.3
django-phonenumber-field==6.1.0
phonenumberslite==8.12.44
Pillow==9.0.1
sqlparse==0.4.2
